requirements.txt
transformers==4.37.2
pillow==10.2.0